Major coverage refactor
Since there are now two coverage maps there was a lot of code duplication.
This breaks up the visualization code up so that more is shared between
the two visualizations.

at some point we may want to run webpack or something on the code to make it
smaller and faster.
