+++ cegs_portal/search/static/search/js/exp_viz/consts.js
@@ -1,0 +1,26 @@
+export const STATE_ZOOMED = "state-zoomed";
+export const STATE_ZOOM_CHROMO_INDEX = "state-zoom-chromo-index";
+export const STATE_SCALE = "state-scale";
+export const STATE_SCALE_X = "state-scale-x";
+export const STATE_SCALE_Y = "state-scale-y";
+export const STATE_VIEWBOX = "state-viewbox";
+export const STATE_FACETS = "state-facets";
+export const STATE_CATEGORICAL_FACET_VALUES = "state-categorical-facets";
+export const STATE_NUMERIC_FACET_VALUES = "state-numeric-facets";
+export const STATE_COUNT_FILTER_VALUES = "state-count-filter";
+export const STATE_COVERAGE_DATA = "state-coverage-data";
+export const STATE_ALL_FILTERED = "state-all-filtered";
+export const STATE_NUMERIC_FILTER_INTERVALS = "state-numeric-filter-intervals";
+export const STATE_COUNT_FILTER_INTERVALS = "state-count-filter-intervals";
+export const STATE_HIGHLIGHT_REGIONS = "state-highlight-regions";
+export const STATE_SELECTED_EXPERIMENTS = "state-selected-experiments";
+export const STATE_SELECTED_EXPERIMENT = "state-selected-experiment";
+export const STATE_ITEM_COUNTS = "state-item-counts";
+export const STATE_SOURCE_TYPE = "state-source-type";
+export const STATE_ANALYSIS = "state-analysis";
+export const STATE_COVERAGE_TYPE = "state-coverage-type";
+export const STATE_LEGEND_INTERVALS = "state-legend-intervals";
+
+export const COVERAGE_TYPE_COUNT = "coverage-type-count";
+export const COVERAGE_TYPE_SIG = "coverage-type-sig";
+export const COVERAGE_TYPE_EFFECT = "coverage-type-effect";
