+++ cegs_portal/search/static/search/js/exp_viz/highlightRegions.js
@@ -1,0 +1,6 @@
+export function getHighlightRegions(regionUploadInput, regionReader) {
+    if (regionUploadInput.files.length != 1) {
+        return;
+    }
+    regionReader.readAsText(regionUploadInput.files[0]);
+}
